deepnsm: sentence-level AriGraph reader — left-corner state machine + P64 discrete substrate

+++ crates/deepnsm/src/quantum_mode.rs
@@ -35,7 +35,7 @@ impl PhaseTag {
         // Use the low 64 bits (the high 64 are reserved for future precision).
         let low = (self.0 & u64::MAX as u128) as u64;
         let normalized = (low as f64) / (u64::MAX as f64);
-        (normalized * std::f64::consts::TAU as f64) as f32
+        (normalized * std::f64::consts::TAU) as f32
     }
 
     pub fn distance(self, other: Self) -> u32 {
@@ -79,7 +79,11 @@ mod tests {
         // f64 intermediate gives sub-1e-3 round-trip; f32 final cast caps
         // precision around 1e-6 of TAU (~6e-6 absolute).
         let diff = (recovered - theta).abs();
-        assert!(diff < 0.001, "round-trip diff {} exceeds tolerance 0.001", diff);
+        assert!(
+            diff < 0.001,
+            "round-trip diff {} exceeds tolerance 0.001",
+            diff
+        );
     }
 
     #[test]
